Fix incompatibility with PostGIS 3.0
It turns out Amazon RDS only supports PostGIS 3.0.x. We were using a variant
of a function that was added in 3.1, which caused a migration failure when
the code was deployed to the staging environment.

Replace the 3.1 syntax with a 3.0-compatible equivalent, and switch to 3.0 for
code generation (which requires migrations to run) so we'll catch any future
problems.

Because this is an edit to an existing migration, it will cause the server to
fail to start up in dev environments that already ran the earlier version.
To fix that, update the checksum manually:

    update flyway_schema_history set checksum = -604226396 where version = '44';