FT-132 Migrate to gis-server species data model (#31)
The species data model supported by gis-server is more sophisticated than the one
we built for the seed bank app in terraware-server. Switch terraware-server to the
better data model.

This is mostly a matter of adding new tables and altering existing ones. However,
there is now a bit of a functional mismatch between the data model and the seed
bank API endpoints; they are designed around the older, less sophisticated data
model, which essentially just treated species and family as strings. The new data
model has a concept of a single "scientific name" per species, and then a list of
arbitrarily many alternate names. In most cases, the existing species names in the
seed bank data are common names rather than scientific ones.

As a temporary fix, we pretend that the species and family names from the seed
bank data set are scientific names. This allows us to maintain backward
compatibility with the existing seed banking apps. Later, once we've updated the
seed banking apps to know about the more sophisticated naming system, someone can
go through any existing species in the database and edit them to move the common
names to the appropriate place and supply the correct scientific names.