FT-133 Add species management endpoints
Add API endpoints to expose the improved species data model to clients. This does
not include endpoints for working with families because the current front end code
doesn't support families.

As part of this change, update the species data model to better reflect how we
want to model names. We want to be able to represent a species whose scientific
name is unknown (so the `species.scientific_name` column in the gis-server schema
isn't suitable) but still require that every species have _some_ kind of canonical
name (so it's not suitable to just remove the species name and rely on child rows
in `species_names`).

The new model has a required species name and a flag to indicate whether it's the
scientific name. We consider that the "primary" name, and its presence is enforced
at the database level. The primary name is also inserted into `species_names` and
the application doesn't allow users to delete it. There can also be additional
names in `species_names`; these can be deleted freely.