FT-180 Remove species merging from seed bank app (#36)
As part of unifying the GIS and seed bank species lists, we want to revisit how we
handle merging species together. Previously, the seed bank API would automatically
merge two species if one of them was renamed to have the same name as the other.
Now the server will return an error if you rename a species and the new name
already exists.