FT-156 Auto-create DAO beans
In the early days of the code base, we were only using a few of the generated
DAO classes so it made sense to only instantiate the specific ones we needed.
But we've been using a bigger and bigger percentage of them over time, so the
burden of having to manually maintain `JooqDaoBeans` is greater than the
benefit of not instantiating unused DAO classes.

Luckily, jOOQ's code generator knows about Spring, so enabling auto-creation
is just a matter of setting a code generator option.