FT-181 Use IAM authentication to connect to database
Add support for using the AWS SDK to generate a temporary authentication token
we can use to connect to an RDS instance. The new behavior is enabled when the
`aws` Spring profile is active. With this change in place, we will no longer
have to put a database password in a config file.