SW-183 Remove per-site config file support (#39)
Initially, we had planned to keep facility configuration information in JSON
files that would be downloaded and cached by terraware-server running on site.
But we've moved away from that model and are going to use the terraware-server
database as the source of truth. There's thus no more need to read the config
file and populate the database. Remove the code for that.