SW-216 Prevent concurrent runs of daily tasks
When we start running multiple terraware-server instances for redundancy, it will
become possible for two or more of them to decide to execute the daily seed bank
tasks at the same time. This could result in duplicate notifications or other bad
side effects.

Add logic to ensure that only one server instance can run a given daily task.

As part of this, move the code that updates the tasks table out of
`TimePeriodTask`, where it never really belonged in the first place. That
interface is now much simpler.

Unit-testing row locking logic is a bit tricky to do reliably since there are no
guarantees about consistent timing in a CI environment. So this includes some
basic functional tests, and I tested by hand by adding `Thread.sleep(30000)` to
various places in the code (e.g., right after the `SELECT FOR UPDATE`, or right
before the `INSERT`) and starting two terraware-server instances to make sure
only one of them ran the tasks.