SW-259 Restrict access to placeholder admin UI
Currently, the placeholder admin UI is accessible to all users. Anyone can
create (and thus become the owner of) a new organization.

This is useful for dev environments and testing, but we don't want random users
stumbling on this ugly UI in production.

Add logic to require the current user to already be an owner or admin before
they can access the admin UI. The idea is that we will create the first admin
user by hand, after which that user can create additional admins. Once we have
a real admin UI, we can get rid of this.

Since the admin UI is useful in dev environments, add a config option to turn
off this new access restriction. Enable the option in the `default` profile so
existing dev environments continue to work as before.