SW-280 Support per-facility seed bank API requests
In the data model, seed data is tied to seed bank facilities. Each accession, and
some other values such as storage locations, has a facility ID so we can track
seeds across multiple seed banks.

Since the UI didn't provide a way to switch between multiple seed banks the user
has access to, as a temporary measure, terraware-server used to assume they only
had access to one: it would look up the list of facilities they had access to and,
if there was just one seed bank entry in that list, use it as the facility ID for
the user's seed bank API calls. If the user had access to more than one seed bank,
it was treated as an error condition.

Modify the API to expose facility IDs to the client.

Since we want accession numbers to only be unique per-facility, this would have
meant making the client pass in a facility ID everywhere it currently uses an
accession number. Instead, modify the API so the client passes in the globally
unique accession ID, from which the server can look up the facility ID.

This requires corresponding changes on the client side.