Suppress spurious warning message from Keycloak adapter
To protect against someone writing malicious serialized Java objects to a token
store, Keycloak added code to restrict the set of classes that could be
deserialized. Unfortunately, that solution breaks down when Keycloak's objects
are deserialized as values in a parent object such as a Spring security
context. In that case Keycloak skips adding the restriction and spits out a
warning message.

We're not using a user-accessible token store, so the security vulnerability is
not really a concern for us. But the warnings are super noisy. Suppress them.