SW-290 Add device configuration endpoints
Add endpoints for creating, updating, and fetching device configuration data.
(There was already one for listing a facility's devices.)

These are intended to be called by API clients for the moment, though at some
point we'll build an admin UI that'll also use them.

The permissions are a little weird here because of the fact that API clients
are created with the "contributor" role. So for now, any operations that the
device manager performs will need to be accessible to contributors. We will
want to revisit this (and there is a task for it already, SW-292).

This also includes a change to make API clients implicit members of all the
projects in their organizations, following the same model we already use for
organization admins, so that they're allowed to access the facilities where
the devices live.