Support non-Keycloak authentication for front end tests (#54)
We want to be able to write front-end tests that don't depend on Keycloak being
available. So we need the tests to be able to authenticate to the server. One
approach is to make a canned login session and have the tests pass its ID in a
cookie just like they would if they'd logged in via Keycloak.

However, it doesn't work to just serialize a real authentication token from
Keycloak, because OIDC tokens always have expiration times and there's a limit
on the maximum lifetime of a token.

So instead, add support for sessions with non-Keycloak authentication data.
There is no way for an end user to create such a session, but for tests, we can
manually create the appropriate objects and insert them into the test database.
A helper method is provided here for just that purpose.