Various GIS endpoint changes (#57)
Add a tag to the Geometry type description so that we can override that type on the client side. This is necessary because the Geometry type is not displayed correctly in the Swagger UI. Therefore, `openapi-typescript` generator is not able to generate the client side type correctly. 

Update `api/v1/gis/plants/list/{layerId}` to return ALL the data from both the Features and Plants tables. Before this change, the endpoint returned some fields from each table.

Update all GIS endpoints to return Geometry coordinates in Longitude/Latitude (EPSG:4326 WGS 84). Before this change, the endpoints accepted Long/Lat coordinates, but returned coordinates in the Spherical Mercator (EPSG:3857 WGS 84) system.