Add total count to feature list response payload (#59)
The front end needs to show the total number of results to the user, so add a
field to the response payload to hold it. For now, the count is calculated with
a naive `SELECT COUNT(*)` query, which should be fine unless we start dealing
with layers that have enormous numbers of features at which point we can
optimize it.