SW-315 Create users from admin UI (#67)
Enhance the "add user to organization" function of the placeholder admin UI to
be able to create users who aren't already registered on the system.

New users get an email message from Keycloak requiring them to set a password.
Existing users get added to the organization.

The "change your password" email redirects the user to the application home
page afterwards. The code attempts to figure out the correct redirect URL.
This works properly in dev environments, but in case it doesn't work in staging
or production thanks to the load balancer, there's a new server config option
to explicitly set the redirect URL.

This isn't intended to be a full-featured "invite user to organization"
implementation, but rather a stopgap to allow us to add alpha users quickly.