SW-315 Increase link validity time to 3 days; minor cleanup (#69)
Feedback from the product team was that the default validity time on "reset
password" links, 12 hours, was too short when they were creating a bunch of
alpha users who might not be at their computers. Increase it to 72 hours as
requested.

Pass the "must set password" setting to Keycloak at user creation time rather
than setting it after creation, to reduce the number of places where a failure
could leave the system in an inconsistent state.

Stop assuming (incorrectly) that the Keycloak admin client `UsersResource.get`
method checks to see if the user exists; it just returns a client object that
can be used to make Keycloak admin API requests related to a specific user.