Redirect after POST in admin UI
Make the POST endpoints in the admin UI redirect to GET endpoints rather than
directly returning an HTML page. This way if the user reloads the page after
submitting a form, the browser won't try to resubmit the form.

This is mostly just a straightforward use of Spring MVC flash attributes (which
are stored in the session data on the server side and get automatically added
to the model and removed from the session store after a redirect) except for
API keys, which we only want to show one time; there's now a GET endpoint to
show a just-created API key, and it redirects to the organization page with an
error message if you try to reload it.