SW-252 Add "Awaiting Check-In" accession state
To support mobile clients uploading accession data before the seeds are delivered
to the seed bank, and to allow the web front end to show a list of such
accessions, add a new accession state that comes before Pending. If an accession
is created in that state, it must be advanced to Pending via an explicit API
request. The existing state transition logic still applies, e.g., if you enter
today's date as the storage date for an Awaiting Check-In accession, it will
transition directly to In Storage.

There's a new optional field in `CreateAccessionRequestPayload` to control
whether an accession starts off in Awaiting Check-In state. The default is the
existing behavior: start off in Pending state (that is, consider new accessions
already checked in at creation time.) Thus any accessions created by existing
clients should continue to behave as expected.

The date and time when an accession was checked in is tracked and is available as
a search field.