SW-326 Centralize permission exceptions
Add a new permission-checking function that throws exceptions if the user doesn't
have the permission in question. Application code like

    if (!user.canCreateSite(projectId)) {
      throw AccessDeniedException("No permission to create site")
    }

changes to

    requirePermissions { createSite(projectId) }

In addition to reducing boilerplate in the application code, this is a first step
toward making the permission checks more intelligent about situations where you
have read-only access to objects. Currently the system treats that situation
inconsistently; sometimes it returns a "not found" error even though you can read
the thing in question. Putting all the exception-throwing code in one place should
make it easier to keep everything consistent.

The KDoc for the new `PermissionRequirements` class lays out some rules for
exception logic which should help people add new code that follows the same
behavior patterns.

However, in this initial version, those rules aren't being followed. This change is
a straight refactor that attempts to exactly reproduce the existing behavior. There
were a couple places where the existing logic was wrong or inconsistent in such a
way that replicating it wouldn't have made sense, but aside from those cases, the
existing unit tests for various permissions continue to pass unmodified.

There is also a unit test for `PermissionRequirements` itself; currently it is not
very interesting but it will get more valuable once we start throwing different
exceptions in different situations.