SW-339 Make accession search result items more strongly typed (#75)
Previously, the OpenAPI schema for the `/api/v1/seedbank/search` response
payload had a `results` property that was just a generic object with string
values and no restrictions on keys. But we know what the possible keys are,
because they can only ever be the names of search fields.

Update the schema to be more precise about the response.

By itself, this doesn't really do the client much good. But when we add nested
values in search results, it will become more valuable because some values will
be objects instead of strings.