SW-252 Add "Awaiting Check-In" accession state (#72)
To support mobile clients uploading accession data before the seeds are delivered
to the seed bank, and to allow the web front end to show a list of such
accessions, add a new accession state that comes before Pending. If an accession
is created in that state, it must be advanced to Pending via an explicit API
request. The existing state transition logic still applies, e.g., if you enter
today's date as the storage date for an Awaiting Check-In accession, it will
transition directly to In Storage.

A new config option `enableAwaitingCheckIn` controls whether or not accessions 
start in Awaiting Check-In state. The default is the existing behavior: start 
off in Pending state (that is, consider new accessions already checked in at
creation time.) Once the front end is updated to know about the new state, the
default can be switched and the code to support the old behavior can be
removed.

The date and time when an accession was checked in is tracked and is available as
a search field.