SW-288 Improve timeseries endpoint error reporting (#76)
`/api/v1/seedbank/timeseries/values` is a batch API endpoint where the client
can submit multiple values for multiple timeseries in one request. Previously,
it just reported that specific values failed to be recorded, but that made it
impossible for the client to tell if the problem was client-side or
server-side.

Add more detailed error messages to differentiate a couple of common problems:
the client specifying a nonexistent timeseries name and the client trying to
record multiple values with the same timestamp.