SW-327 Move all parent ID lookups into permission check methods (#78)
Previously, as an optimization, we had methods in `UserModel` that checked for
permission to do something with a child object but took the ID of a parent
object as a parameter, since we knew the permissions were always inherited.
But this makes it tough to throw the correct `EntityNotFoundException` because
it should refer to the thing the user was actually trying to access, which is
the child object. To work around that, in some places the methods took both
the parent and child object IDs, forcing calling code to know about the
permission boundaries. But sometimes the methods just took the child IDs and
looked the parent IDs up. The whole thing was inconsistent.

Update the `UserModel` and `PermissionRequirements` APIs so they only ever take
the ID of the object the user is trying to access. They now always look up
parent IDs as needed.

As a side benefit, this means we can now declare that the permission checking
methods will always do an existence check. Previously that was _usually_ the
case but not always: if a method took a parent ID as an argument, it could
pass the permission check even if the child didn't exist. Making that behavior
official means we can get rid of some existence checks in the calling code,
making this change performance-neutral in those places. (It still causes
extra database queries in other places, though.) There's a basic test case to
sanity-check that permission checks always fail on nonexistent objects.