SW-291 Use consistent alignment of API payload field declarations
If both an annotation and a variable declaration fit on the same line, ktfmt
puts them both on one line. But it makes that decision on a per-variable basis
which means if you have a bunch of annotated variables in a row, some of them
will be line-wrapped and some won't. This is hard to visually scan.

In our code base, this is almost exclusively a problem in API payload classes
with the `@Schema` annotation.

Force ktfmt to wrap these declarations onto multiple lines by putting a
trailing comma in the annotation's parameter list. Parameter lists with
trailing commas are always split into multiple lines.

There are no functional changes here, just code formatting.