SW-278 Add API endpoints for facility automations
The device manager will support various automations that can be run at a
facility. Each automation needs automation-specific configuration data. Add a
series of CRUD endpoints to manage the data.

The automations are stored in a new `automations` table. The Flyway script that
creates that table introduces something we should have been using all along:
instead of having a separate SQL file for Postgres and non-Postgres databases
where the only difference is the type of the JSONB column, instead it uses
Flyway's built-in variable substitution feature and defines a variable for
the column type that should be used for JSON data.