SW-279 Add device manager notification endpoint
Add an API endpoint the device manager can use to send email notifications when it
detects a problem at a facility. This is a stopgap until we have a more complete
design for how alerting and monitoring should work.

The notifications are sent to a per-facility list of email addresses. The admin UI
now has a facility page (previously, facilities were just listed on the site page)
where the notification addresses can be edited. Other facility information can be
edited there too.

The server can be configured to send mail using either SMTP or the AWS SES API.
A series of config options can be set for use in dev or staging environments to
control the recipients and subject lines of email messages. By default, it just
writes messages to the server log and doesn't try to send them at all.