Generate GitHub Pages site with schema documentation
Add a GitHub Actions workflow to automatically generate database schema
documentation and publish it to the `gh-pages` branch, which GitHub will then turn
into a web site people can use to browse the data model. Docs are generated
whenever there's a change to the database migrations.

The schema documentation is generated using SchemaSpy, which we can configure
using a generator class that runs as part of the test suite. The actual
documentation generation is disabled by default since it can take a couple minutes
to run, but the test class also sanity-checks that the docs are properly
configured, and those tests run as part of the normal suite.

As part of generating the docs, add comments to all the tables in our database.
The comments are now all in a single SQL script rather than scattered around in
individual migrations. Existing comments from older migrations have been copied to
the new script.