TRACK-379 Fix SES sender address
Now that SES is set up on our non-production account, I'm able to actually test
the SES code path, and there was a bug: the code was looking in the wrong place
for the sender address to include in the email sending API call. Fix it to look
in the right place.

Our AWS environments aren't able to send email to the public Internet yet, but
I tested this by sending to the SES test mailbox address and confirming the
SES API calls succeeded.