SW-339 Support nested search results (#84)
Rewrite the code that generates SQL queries to search for accessions. The new
implementation supports nested lists of values in search results. These
correspond to the field structure in the accession payload: there is a list
of germination tests each of which contains a list of germinations, etc.

The existing flat-table Cartesian join representation of lists of values is
still supported and can be freely mixed and matched with nested fields.
Users of the existing front end shouldn't notice any behavior difference;
this was confirmed using the web app test suite.

Currently, nested fields aren't supported for CSV export, simply because it's
unclear how one should format a list of lists of values in a CSV cell.

The implementation leans heavily on jOOQ multisets to bind nested values to
search results. It works around some limitations of multisets, which makes the
code kind of complicated. `NestedQueryBuilder.kt` has what is hopefully a
sufficiently thorough description of why it needs to work the way it works.

Although the code as it exists here has a few hardwired references to the
accessions table, it's designed with an eye toward making it easy to modify
to work with other kinds of data in the future if we want similar search
functionality.