Remove explicit ID lists from PermissionTest (#90)
Previously, the assertions in `PermissionTest` checked a bunch of hardwired
object IDs for the correct permissions. This was repetitive and hard to read
because it wasn't immediately obvious what a given list of IDs was supposed to
represent.

Extract those ID lists into variables. This reduces repetition and the variable
names should help make it clearer what behavior the tests are trying to verify.