TRACK-418 Start decoupling search code from seed bank tables (#93)
We want to support searching across GIS and organization data. The recent
changes to the search code were made with that in mind, but they were still
tied to the seed tables to some degree.

As a first step in decoupling, move a bunch of interfaces and code into a new
`com.terraformation.backend.search` package. The idea is that that package will
have the search framework, and then data-type-specific classes will have the
logic that is specific to a particular search topic.

There are two real changes here:

1. `SearchTable.leftJoinWithAccession` is now `SearchTable.leftJoinWithMain`.
2. `NestedQueryBuilder.getSearchTable` now computes the root node's table
   based on the list of fields, rather than hardwiring it to `accessions`.

Other than those two things, everything else is just moving code into the new
package without changing anything.