TRACK-419 Add plant details to features API
Update the API payload classes for features to add plant information in the
form of an optional child object. Clients can include plant information when
they create features, add it to existing features, and retrieve it as part of
the feature data.

The plants API continues to be supported for backward compatibility, but its
functionality should now be covered by the features API. We can remove it once
clients are updated.

As part of updating the features API, I also cleaned up a couple minor things:

* There's no longer a need to catch `FeatureNotFoundException` and rethrow it
  as `NotFoundException`.
* Clients no longer have to pass in a layer ID when updating a feature.
  Previously, it was required but wasn't modifiable.