TRACK-419 Merge plants and features internally (#95)
Change the internal modeling of plants so they are treated as optional
supplementary data about features, rather than as first-class entities. As we
add other feature types, there is now a straightforward way to model them in
the application code.

This is purely an implementation change; the client-facing API continues to behave
exactly as before.

With this structure, it no longer makes sense for plants to have their own
creation and modification timestamps, so drop those columns from the plants table.

`PlantStore` is now merged into `FeatureStore`. It still has methods to directly
operate on plants, though those will go away once we no longer need to support the
plant-centric API endpoints.

Likewise, `PlantStoreTest` is rolled into `FeatureStoreTest`, minus test cases
related to plants having timestamps.

As part of this change, I also fixed two bugs in `fetchPlantsList`: it wasn't
filtering by layer ID, and it was returning search results in an undefined order.
The latter issue caused test cases to fail if the database didn't happen to return
rows in insertion order.