Allow passing ID wrappers to test data setup methods
The test data setup methods in `DatabaseTest` take `Long` arguments. That
helps make tests succinct when ID values are specified explicitly (you can do
`insertOrganization(1)` and so on) but if we already have a wrapped ID, it
means we have to unwrap it before we can insert it.

Modify those methods to accept either wrapped IDs or raw numbers. Stop
unwrapping IDs in tests where we needed to do so before.