TRACK-418 Decouple search fields from seed tables (#94)
Continuing the groundwork for search APIs for organization and GIS data, move the
search field classes out of the seedbank package.

Since the classes are no longer inner classes of the seedbank `SearchFields` class,
they can no longer refer to that class's `fuzzySeachOperators` and instead have to
take it as an argument, which would be annoying to have to do for every text field
because there are a lot of them.

It also no longer makes sense for the `accessions` search table to be the default.
But it would still be pretty annoying to have to pass it to their constructors
a zillion times in `SearchFields`.

To solve both those problems, `SearchTable` is now an abstract class instead of an
interface, and it has a bunch of helper methods to create fields. The net effect
is that in `SearchFields`, instead of doing

    TextField(
            "fieldName",
            "displayName",
            GERMINATIONS.WHATEVER,
            SearchTables.Germination)

you now do

    searchTables.germinations.textField(
            fieldName",
            "displayName",
            GERMINATIONS.WHATEVER)

There are helper methods for all the field types we use in the existing seedbank
search API, except for a handful such as the `active` field that are very specific
to the seed data model and wouldn't be useful elsewhere. Those fields continue to
be implemented as one-off inner classes in `SearchFields`.

The search table definitions in the seedbank `SearchTables` class are now `val`s
instead of static `object`s because they need to take constructor arguments that
aren't available until dependency injection happens.

With these changes, it should be possible to define a set of search fields and
tables that have nothing to do with the seedbank data model.