TRACK-465 Richer data model for search field paths
Rework the way the code represents search field names internally. Previously, they
were represented as strings, and the code used string manipulation to navigate
them while constructing the tree of sublists and scalar fields. This was
inflexible and would have made it awkward to support a more flexible search API
that can navigate the data model more freely.

The client-facing API still uses strings, but now those strings are parsed into
path objects where each element of the path carries information about what part of
the data model it refers to. The list of field definitions is now broken apart
into separate "namespaces" that are arranged in a hierarchy, rather than defined
as a flat list where the hierarchy of sublists was implicit in the field names.

There is no functionality change here! This is a pure refactor and the accession
search API should behave identically from the client's point of view. The existing
test suite is unmodified here aside from switching to a different method for
resolving field names.

Subsequent changes will build on this to start building a more sophisticated
client-facing API.