TRACK-428 Fetch organization hierarchy (#98)
Update the organization list endpoint to optionally also return a list of 
projects, sites, and facilities for each organization. The client can specify
how deep it wants the results to be.

Add an endpoint to get the information for a single organization as opposed to
listing all of the available ones.